Starting to flesh out GameController
This is going to take a while. GameController as per issue #1 should be
a static class capable of processing game requests and sending them to a
specific game. What I have in mind is some sort of system for numbering
games in an external store (either database or known memory structure)
and then accessing the game entirely through passing "messages" to that,
with the game knowing how to respond to messages and the controller
knowing what is expected (on a send, nothing; on a retrieve, some out
message, etc). Please comment on this commit if you have thoughts.