Scrapping old server code and including basic outline for new stuff.
In regards to issue #1, we are detangling the code for the server from
the code for the game by moving game elements, visualizer elements, and
player elements out into seperate controllers. What remains now are
stubs for working on the HTTP request handler so it can process incoming
data and pass it along to the correct controller.