Fixing broken testing suite and adding commit hook, resolves #5
Removed the retarded random_tests file that fails by default, fixed
test_suite so that it can be run externally and returns the results to
the external runner (test_suite should be further fixed so that lines
don't need to be added every time someone introduces a new test system),
and created a pre-commit hook in Python that checks test cases on core
branches and blocks committing to main branches. All in all, some pretty
sweet work done here.