Fuck GET, we're doing POST live
Stupid oversight that the client can't actually pass a JSON dictionary
with an id/auth pair over a GET request. We weren't RESTful anyways
(since we didn't have URLs which supported multiple request types), so
now we're just making everything a POST request. Probably not the best
solution, but with the time crunch this really works fine.