Implemented a basic arena.
This script expects there to be exactly four teams, every time. It will
take four team names and run them against each other. It looks for
folders with that team's name and a client file (binary in any language)
and executes it, passing in a command line argument which is the full
address to the server the game is running on (server name and port).